Refactor DetailQuizView
