+++ 4GQuizApp/Module/DetailQuiz/View/DetailQuizView.swift
@@ -8,7 +8,7 @@
 import SwiftUI
 
 struct DetailQuizView: View {
-    @StateObject private var viewModel = DetailQuizViewModel()
+    @StateObject var viewModel = DetailQuizViewModel()
     @Environment(\.dismiss) private var dismissView
     @State private var showScoreCardView: Bool = false
     @State private var showPopupInfo: Bool = false
@@ -156,72 +156,5 @@ struct DetailQuizView: View {
                 .hCenter()
             })
         }
-    }
-    
-    @ViewBuilder
-    func QuestionView(_ question: Question)-> some View {
-        VStack(alignment: .leading, spacing: 5) {
-            if !(question.image.width?.isEmpty ?? false) {
-                CustomImageHelper(imagekey: ImageHelper.shared.getPath(width: Int(question.image.width ?? "") ?? 0,
-                                                                       height: Int(question.image.height ?? "") ?? 0,
-                                                                       urlName: question.image.url))
-                .frame(height: 200)
-                .cornerRadius(12, corners: .allCorners)
-            }
-            
-            Text("Pytania: \((viewModel.currentIndex) + 1)/\(quiz.questions.count)")
-                .font(.callout)
-                .foregroundColor(.gray)
-                .hLeading()
-            
-            Text(question.text)
-                .headerDetailTitle()
-                .frame(height: CGFloat(question.text.count))
-                .lineLimit(5)
-            
-            VStack(spacing: 12) {
-                ForEach(question.answers,id:\.order) { option in
-                    ZStack {
-                        AnswerView(option.text, tint: .gray)
-                            .opacity(viewModel.isSelectedAnswer == option.text && viewModel.isSelectedAnswer != "" ? 0 : 1)
-                        AnswerView(option.text, tint: .royal)
-                            .opacity(viewModel.isSelectedAnswer == option.text && viewModel.isSelectedAnswer != "" ? 1 : 0)
-                    }
-                    .contentShape(Rectangle())
-                    .onTapGesture {
-                        self.viewModel.isSelectedAnswer = option.text
-                        self.viewModel.answer = option
-                    }
-                }
-            }
-            .padding(.vertical,10)
-        }
-        .padding(15)
-        .hCenter()
-        .background {
-            RoundedRectangle(cornerRadius: 20, style: .continuous)
-                .fill(.white)
-        }
-        .padding(.horizontal,15)
-    }
-    
-    @ViewBuilder
-    func AnswerView(_ answer: String, tint: Color)-> some View {
-        Text(answer)
-            .font(.system(size: 12, weight: .regular))
-            .foregroundColor(tint)
-            .padding(.horizontal,15)
-            .padding(.vertical,10)
-            .lineLimit(2)
-            .frame(height: 50)
-            .hLeading()
-            .background {
-                RoundedRectangle(cornerRadius: 12, style: .continuous)
-                    .fill(tint.opacity(0.15))
-                    .background {
-                        RoundedRectangle(cornerRadius: 12, style: .continuous)
-                            .stroke(tint.opacity(tint == .gray ? 0.15 : 1), lineWidth: 2)
-                    }
-            }
     }
 }
